feat(engine): add executed decision elements metrics
related to #CAM-4536

+++ engine/src/main/java/org/camunda/bpm/engine/impl/metrics/Meter.java
@@ -34,7 +34,7 @@ public class Meter {
     counter.incrementAndGet();
   }
 
-  public void markTimes(int times) {
+  public void markTimes(long times) {
     counter.addAndGet(times);
   }
